+++ ByteCat/src/ByteCat/app/Window.h
@@ -2,7 +2,6 @@
 
 #include <string>
 #include "byteCat/input/events/Event.h"
-#include "glm/vec2.hpp"
 
 namespace BC
 {
@@ -28,7 +27,7 @@ namespace BC
 		void update() const;
 		void shutdown() const;
 
-		void resize(float x, float y);
+		void resize(unsigned int x, unsigned int y);
 		
 		std::string getTitle() const { return setting.title; }
 		int getWidth() const { return setting.width; }
